// web/src/types.ts
export interface Exchange {
  id: string // UUID (empty for supported exchange templates)
  exchange_type: string // "binance"
  account_name: string // User-defined account name
  name: string // Display name
  type: 'cex' | 'dex'
  enabled: boolean
  apiKey?: string
  secretKey?: string
  testnet?: boolean
  hyperliquidWalletAddr?: string
  asterUser?: string
  asterSigner?: string
  asterPrivateKey?: string
  lighterWalletAddr?: string
  lighterPrivateKey?: string
  lighterApiKeyPrivateKey?: string
  lighterApiKeyIndex?: number
}